# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DatasetAvailability": {
        "description": "What the interactive views can render for a run, and why not when they can't.\n\nThe run detail view calls this first to choose a layout, so it answers for\nevery run -- including angular and pre-contract ones -- instead of erroring.",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/DatasetKind"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Human-readable explanation, safe to show",
            "title": "Message"
          },
          "profiles_available": {
            "default": false,
            "title": "Profiles Available",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UnavailableReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "sigmas_available": {
            "default": false,
            "description": "Whether uncertainties exist; calc_sigmas is off by default",
            "title": "Sigmas Available",
            "type": "boolean"
          },
          "spectra": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SpectrumInfo"
            },
            "title": "Spectra",
            "type": "array"
          },
          "supported": {
            "description": "True when at least one spectrum can be served",
            "title": "Supported",
            "type": "boolean"
          },
          "unavailable_fields": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Fields a client might expect but which cannot be served, mapped to why",
            "title": "Unavailable Fields",
            "type": "object"
          }
        },
        "required": [
          "kind",
          "supported"
        ],
        "title": "DatasetAvailability",
        "type": "object"
      },
      "DatasetKind": {
        "description": "What kind of Thomson run the artifacts say this is.",
        "enum": [
          "one_d",
          "angular",
          "unknown"
        ],
        "title": "DatasetKind",
        "type": "string"
      },
      "DatasetUnavailableBody": {
        "description": "The reason payload itself.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/UnavailableReason"
          }
        },
        "required": [
          "reason",
          "detail"
        ],
        "title": "DatasetUnavailableBody",
        "type": "object"
      },
      "DatasetUnavailableResponse": {
        "description": "Error body for the dataset endpoints, carrying a reason code.\n\nNested under ``detail`` because that is what actually goes over the wire:\nFastAPI wraps whatever an ``HTTPException`` carries in a ``detail`` key. A\nflat declaration would make the generated client read ``err.reason`` and get\n``undefined`` on exactly the angular-vs-missing distinction these endpoints\nexist to make legible.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/DatasetUnavailableBody"
          }
        },
        "required": [
          "detail"
        ],
        "title": "DatasetUnavailableResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Lineout": {
        "properties": {
          "components": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "description": "Always empty today; see components_unavailable",
            "title": "Components",
            "type": "object"
          },
          "components_unavailable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why IRF/noise components are absent, when they are",
            "title": "Components Unavailable"
          },
          "data": {
            "description": "Measured spectrum",
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Data",
            "type": "array"
          },
          "fit": {
            "description": "Fitted spectrum",
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Fit",
            "type": "array"
          },
          "index": {
            "description": "Resolved lineout index; negative inputs are normalized",
            "title": "Index",
            "type": "integer"
          },
          "lineout_count": {
            "title": "Lineout Count",
            "type": "integer"
          },
          "residual": {
            "description": "Derived as data - fit",
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Residual",
            "type": "array"
          },
          "wavelength": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Wavelength",
            "type": "array"
          },
          "which": {
            "title": "Which",
            "type": "string"
          },
          "x_label": {
            "title": "X Label",
            "type": "string"
          },
          "x_value": {
            "description": "Position of this lineout on the lineout axis",
            "title": "X Value",
            "type": "number"
          },
          "y_label": {
            "title": "Y Label",
            "type": "string"
          }
        },
        "required": [
          "which",
          "index",
          "lineout_count",
          "x_label",
          "x_value",
          "y_label",
          "wavelength",
          "data",
          "fit",
          "residual"
        ],
        "title": "Lineout",
        "type": "object"
      },
...
      "ProfileSeries": {
        "properties": {
          "name": {
            "description": "Column name from learned_parameters.csv, e.g. 'Te_electron'",
            "title": "Name",
            "type": "string"
          },
          "sigma": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-point uncertainty when the run computed sigmas",
            "title": "Sigma"
          },
          "values": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "name",
          "values"
        ],
        "title": "ProfileSeries",
        "type": "object"
      },
      "Profiles": {
        "properties": {
          "lineout_pixels": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Detector pixel index per lineout, when logged",
            "title": "Lineout Pixels"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/ProfileSeries"
            },
            "title": "Series",
            "type": "array"
          },
          "sigmas_available": {
            "title": "Sigmas Available",
            "type": "boolean"
          },
          "x": {
            "description": "Lineout axis coordinates",
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "X",
            "type": "array"
          },
          "x_label": {
            "title": "X Label",
            "type": "string"
          }
        },
        "required": [
          "x_label",
          "x",
          "series",
          "sigmas_available"
        ],
        "title": "Profiles",
        "type": "object"
      },
...
      "Spectrogram": {
        "properties": {
          "downsample_factors": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Block-averaging factors applied per axis. The lineout axis is spared where possible.",
            "title": "Downsample Factors",
            "type": "object"
          },
          "downsample_method": {
            "description": "'mean' when downsampled, 'none' when full resolution",
            "title": "Downsample Method",
            "type": "string"
          },
          "field": {
            "description": "'data', 'fit', or the derived 'residual'",
            "title": "Field",
            "type": "string"
          },
          "full_shape": {
            "description": "[lineouts, wavelengths] before downsampling",
            "items": {
              "type": "integer"
            },
            "title": "Full Shape",
            "type": "array"
          },
          "returned_shape": {
            "description": "[lineouts, wavelengths] as returned",
            "items": {
              "type": "integer"
            },
            "title": "Returned Shape",
            "type": "array"
          },
          "values": {
            "description": "Shape (len(y), len(x)) -- row-major by wavelength, directly usable as a Plotly heatmap z",
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Values",
            "type": "array"
          },
          "which": {
            "title": "Which",
            "type": "string"
          },
          "x": {
            "description": "Lineout axis coordinates",
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "X",
            "type": "array"
          },
          "x_label": {
            "title": "X Label",
            "type": "string"
          },
          "y": {
            "description": "Wavelength coordinates",
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Y",
            "type": "array"
          },
          "y_label": {
            "title": "Y Label",
            "type": "string"
          }
        },
        "required": [
          "which",
          "field",
          "x_label",
          "y_label",
          "x",
          "y",
          "values",
          "full_shape",
          "returned_shape",
          "downsample_factors",
          "downsample_method"
        ],
        "title": "Spectrogram",
        "type": "object"
      },
      "SpectrumInfo": {
        "properties": {
          "fields": {
            "description": "Renderable fields; 'residual' is derived as data - fit",
            "items": {
              "type": "string"
            },
            "title": "Fields",
            "type": "array"
          },
          "lineout_count": {
            "title": "Lineout Count",
            "type": "integer"
          },
          "path": {
            "description": "Artifact path this spectrum was read from",
            "title": "Path",
            "type": "string"
          },
          "wavelength_count": {
            "title": "Wavelength Count",
            "type": "integer"
          },
          "which": {
            "description": "'ele' or 'ion'",
            "title": "Which",
            "type": "string"
          },
          "x_label": {
            "description": "Lineout axis name, e.g. 'Time (ps)' or 'Radius (\\mum)'",
            "title": "X Label",
            "type": "string"
          },
          "y_label": {
            "description": "Spectral axis name, normally 'Wavelength'",
            "title": "Y Label",
            "type": "string"
          }
        },
        "required": [
          "which",
          "path",
          "x_label",
          "y_label",
          "lineout_count",
          "wavelength_count",
          "fields"
        ],
        "title": "SpectrumInfo",
        "type": "object"
      },
...
      "UnavailableReason": {
        "description": "Machine-readable reason a dataset view cannot be served.\n\nDistinguishing these matters: showing \"no data found\" for an angular run\nreads as a bug, when the truth is that the view is deliberately out of scope.",
        "enum": [
          "angular_not_supported",
          "dataset_missing",
          "dataset_unreadable",
          "unexpected_schema",
          "field_unavailable",
          "index_out_of_range"
        ],
        "title": "UnavailableReason",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nRead layer over the MLflow tracking server for browsing Thomson scattering\nanalysis runs. MLflow remains the source of truth; this service holds no\ndatabase of its own and never writes to it.\n",
    "title": "TSADAR analysis browser API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs/{run_id}/datasets": {
      "get": {
        "description": "Probe before rendering.\n\nAlways answers -- an angular or pre-contract run gets ``supported: false``\nplus a reason rather than an error, so the run detail view can pick a layout\nin one call.",
        "operationId": "get_availability_api_runs__run_id__datasets_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetAvailability"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "What interactive views this run supports",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/runs/{run_id}/lineout": {
      "get": {
        "description": "The interactive replacement for the pre-rendered lineout PNGs.\n\nIRF and noise components are not in the netCDF datasets, so they are reported\nas unavailable rather than fabricated -- see ``components_unavailable``.",
        "operationId": "get_lineout_api_runs__run_id__lineout_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "'ele' or 'ion'",
            "in": "query",
            "name": "which",
            "required": false,
            "schema": {
              "default": "ele",
              "description": "'ele' or 'ion'",
              "title": "Which",
              "type": "string"
            }
          },
          {
            "description": "Lineout index; negative counts from the end",
            "in": "query",
            "name": "index",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Lineout index; negative counts from the end",
              "title": "Index",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lineout"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetUnavailableResponse"
                }
              }
            },
            "description": "Invalid field, spectrum or index"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetUnavailableResponse"
                }
              }
            },
            "description": "Dataset missing or unreadable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetUnavailableResponse"
                }
              }
            },
            "description": "Recognized but unsupported, e.g. an angular run"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Measured vs fitted spectrum at one lineout",
        "tags": [
          "datasets"
        ]
      }
    },
...
    "/api/runs/{run_id}/profiles": {
      "get": {
        "operationId": "get_profiles_api_runs__run_id__profiles_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profiles"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetUnavailableResponse"
                }
              }
            },
            "description": "Invalid field, spectrum or index"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetUnavailableResponse"
                }
              }
            },
            "description": "Dataset missing or unreadable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetUnavailableResponse"
                }
              }
            },
            "description": "Recognized but unsupported, e.g. an angular run"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fitted parameters vs lineout, with uncertainties where available",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/runs/{run_id}/spectrogram": {
      "get": {
        "operationId": "get_spectrogram_api_runs__run_id__spectrogram_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "'ele' or 'ion'",
            "in": "query",
            "name": "which",
            "required": false,
            "schema": {
              "default": "ele",
              "description": "'ele' or 'ion'",
              "title": "Which",
              "type": "string"
            }
          },
          {
            "description": "One of ['data', 'fit', 'residual']; 'residual' is derived as data - fit",
            "in": "query",
            "name": "field",
            "required": false,
            "schema": {
              "default": "data",
              "description": "One of ['data', 'fit', 'residual']; 'residual' is derived as data - fit",
              "title": "Field",
              "type": "string"
            }
          },
          {
            "description": "Pixel budget. The array is block-averaged to at most this many points; the lineout axis is spared where possible, so wavelength is reduced first.",
            "in": "query",
            "name": "max_px",
            "required": false,
            "schema": {
              "default": 20000,
              "description": "Pixel budget. The array is block-averaged to at most this many points; the lineout axis is spared where possible, so wavelength is reduced first.",
              "minimum": 1,
              "title": "Max Px",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spectrogram"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetUnavailableResponse"
                }
              }
            },
            "description": "Invalid field, spectrum or index"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetUnavailableResponse"
                }
              }
            },
            "description": "Dataset missing or unreadable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetUnavailableResponse"
                }
              }
            },
            "description": "Recognized but unsupported, e.g. an angular run"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "2D spectrogram, downsampled server-side",
        "tags": [
          "datasets"
        ]
      }
    }
  }
}